Revert "[csswg] Maintain the hash on redirect."
This reverts commit 53bed3c85985d5cd39973389667b54c0611181a9.

It seems that without this change, the behavior is already such that the
hash would be maintained across the redirects — but with this change,
https://w3c.github.io/csswg-drafts/css-backgrounds/#the-background redirects
to https://drafts.csswg.org/css-backgrounds/#the-background##the-background
(that is, the hash part gets repeated, and with a doubled ## too).

Given that, it seems that completely reverting this change should get
the redirects working as expected.