"Implementation Report" no longer required at CR steps
While it remains good practice to have a preliminary implementation report when
a spec moves to Candidate Recommendation, "a statement that no such report
exists" is ok from a publication rules perspective:
https://www.w3.org/pubrules/doc/rules/?profile=CR#implReport

Bikeshed required a link to an implementation report at CR phases. This update
drops the requirement to cater for specs (such as WebGPU) that make a statement
in the Status of this Document section.